requirements.txt
questionary==1.11.0
pyyaml==6.0
torch==2.1.0
torchvision==0.16.0
